+++ src/dcos_e2e/backends/_base_classes.py
@@ -41,7 +41,7 @@ class ClusterManager(abc.ABC):
     def install_dcos_from_url(
         self,
         build_artifact: str,
-        extra_config: Dict[str, Any],
+        dcos_config: Dict[str, Any],
         log_output_live: bool,
     ) -> None:
         """
@@ -50,9 +50,7 @@ class ClusterManager(abc.ABC):
         Args:
             build_artifact: The URL string to a build artifact to install DC/OS
                 from.
-            extra_config: This may contain extra installation configuration
-                variables that are applied on top of the default DC/OS
-                configuration for a particular backend.
+            dcos_config: The DC/OS configuration to use.
             log_output_live: If ``True``, log output of the installation live.
         """
 
@@ -60,7 +58,7 @@ class ClusterManager(abc.ABC):
     def install_dcos_from_path(
         self,
         build_artifact: Path,
-        extra_config: Dict[str, Any],
+        dcos_config: Dict[str, Any],
         log_output_live: bool,
     ) -> None:
         """
@@ -68,9 +66,7 @@ class ClusterManager(abc.ABC):
 
         Args:
             build_artifact: The path to a build artifact to install DC/OS from.
-            extra_config: May contain extra installation configuration
-                variables that are applied on top of the default DC/OS
-                configuration for a particular backend.
+            dcos_config: The DC/OS configuration to use.
             log_output_live: If ``True``, log output of the installation live.
         """