+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ Next
 ----
 
 - Show a formatted error rather than a traceback if Docker cannot be connected to.
+- Custom backends' must now implement a ``base_config`` method.
+- Custom backends' installation methods must now take ``dcos_config`` rather than ``extra_config``.
 
 2018.05.17.0
 ------------